+++ tests/CustomCasterTest.php
@@ -2,20 +2,20 @@
 
 namespace Glhd\LaravelDumper\Tests;
 
-use Glhd\LaravelDumper\Facades\LaravelDumper;
+use Glhd\LaravelDumper\Casters\CustomCaster;
 
 class CustomCasterTest extends TestCase
 {
 	public function test_custom_caster(): void
 	{
-		LaravelDumper::for(MyCustomObject::class)
+		CustomCaster::for(MyCustomObject::class)
 			->only(['foo', 'nothing', 'nah'])
 			->dynamic('dyn', fn() => 'this is a dynamic prop')
 			->virtual('virt', fn() => 'this is a virtual prop')
 			->filter()
 			->reorder(['dyn', 'foo']);
 		
-		LaravelDumper::for(MyOtherCustomObject::class)
+		CustomCaster::for(MyOtherCustomObject::class)
 			->virtual('foo', fn() => 'bar');
 		
 		$expected = <<<EOD